training_scripts/requirements.txt
# image already installed
# torch==2.0.1
# deepspeed==0.11.1

# current images env
# transformers==4.38.1
# accelerate==0.29.2
# torch==2.1.0
# deepspeed==0.10.2
# tokenizers==0.15.2

h5py
numpy
rouge_score
fire
openai==0.28.1
# transformers>=4.35.2
# accelerate>=0.22.0
sentencepiece
# tokenizers>=0.13.3
jsonlines
Pillow
protobuf